Spring kotlin setup

+++ SpringLexiLucha/src/main/kotlin/com/LexiLucha/LexiLucha/model/CompletedQuestion.kt
@@ -1,4 +1,4 @@
 package com.LexiLucha.LexiLucha.model
 
-data class CompletedQuestion(val questionId: Int, val timeTaken: Int, val correct: Boolean = true) {
+data class CompletedQuestion(val questionId: Int, val timeTaken: Long, val correct: Boolean = true) {
 }
